% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_si_gt.R
\name{compute_gt}
\alias{compute_gt}
\title{Compute generation time values}
\usage{
compute_gt(df)
}
\arguments{
\item{df}{The input data frame.}
}
\value{
The modified data frame with the added gt column.
}
\description{
Compute the generation time values for the transmission pairs.
}
\keyword{internal}
