% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_si_gt.R
\name{compute_si}
\alias{compute_si}
\title{Compute serial interval values}
\usage{
compute_si(df)
}
\arguments{
\item{df}{The input data frame.}
}
\value{
The modified data frame with the added si column.
}
\description{
Compute the serial interval for the transmission pairs.
}
\keyword{internal}
