% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_labels.R
\name{draw_labels}
\alias{draw_labels}
\title{Genrate random labels}
\usage{
draw_labels(n = 1, size = 6)
}
\arguments{
\item{n}{the number of labels to generate; defaults to 1}

\item{size}{the number of characters of the labels; defaults to 6}
}
\description{
This function is a simple wrapper for \code{stri_rand_strings} from the \code{stringi}
package to generate random alphanumeric labels.
}
\examples{

draw_labels()
draw_labels(10, 3)
}
