% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_methods.R
\name{as_epicontacts}
\alias{as_epicontacts}
\alias{outbreak}
\alias{outbreak_class}
\alias{as.epicontacts}
\alias{as_epicontacts.outbreaks}
\alias{as.epicontacts.outbreaks}
\alias{as_epicontacts.outbreak}
\alias{plot.outbreak}
\title{Basic methods for outbreaks objects}
\usage{
as_epicontacts(...)

as.epicontacts(...)

\method{as_epicontacts}{outbreak}(x, ...)

\method{plot}{outbreak}(x, ...)
}
\arguments{
\item{...}{for \code{plot}, arguments passed to \code{plot.epicontacts} in the
\code{epicontacts} package.}

\item{x}{an object to be converted to \code{epicontacts}}
}
\description{
These functions implement basic methods for outbreaks objects. They include:
\itemize{
\item \code{as_epicontacts}: conversion to \code{epicontacts} object
\item \code{as.epicontacts}: alias for \code{as_epicontacts}
}
}
