% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_contacts.R
\name{simulate_contacts}
\alias{simulate_contacts}
\title{Simulate contacts of cases}
\usage{
simulate_contacts(x, n_contacts, dist_time_to_contact, dist_duration)
}
\arguments{
\item{x}{an \code{outbreak} object as returned by \code{simulate_outbreak}}

\item{n_contacts}{a vector of values to be used as average number of
contacts; values will be drawn at random from this vector to determine the
expected number of contacts for each new case}

\item{dist_time_to_contact}{the distribution of the delay to new exposures,
in days after the onset of symptoms}

\item{dist_duration}{the distribution of duration of the exposure period
after the first day, i.e. the time interval between the start and the end
of the exposure window}
}
\description{
Under development. Do not use. This simulator adds contacts to the cases of
an \code{outbreak} object, using user-defined distributions for numbers and
duration of contacts. It returns an \code{outbreaks} object with cases and
contacts, identifying entries in the linelist as cases or contacts, and links
as exposures or transmissions.
}
\examples{

## first simulate an outbreak
## (see example in ?simulate_outbreak)
incubation <- c(0, 1, 1, 1, 1) # numbers = unscaled PMF
infectious_period <- make_disc_gamma(10, 7) # distcrete object
reporting <- function(x) dpois(x, 5) # PMF function
set.seed(1)
x <- simulate_outbreak(R = runif(100, 1, 3), # random values on [1;3]
                       dist_incubation= incubation,
                       dist_infectious_period = infectious_period,
                       dist_reporting = reporting)

## simulate contacts: inputs

## exposure starts 0-2 days post onset
time_to_contact = c(1, 1, 1)

## geom dist for duration of exposure
duration <- function(x) dgeom(x, prob = .9)

x_with_contacts <- simulate_contacts(
    x[1:10, ],
    n_contacts = 1:10, # 1 to 10 contacts
    dist_time_to_contact = time_to_contact,
    dist_duration = duration)

## check output
class(x_with_contacts)
dim(x_with_contacts)
head(x_with_contacts)

plot(x_with_contacts)

}
\seealso{
\code{simulate_outbreak}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
